LuaChat First Commit
This repository contains a sample application to demonstrate
how to combine C++17 and Lua 5.3 (using SWIG and CMake). LuaChat
allows two users to connect (via TCP) and to exchange messages in
a similar way to the Unix 'talk' program.

Each commit documents a key step in the code's development. This
commit provides enough code for the repository to be built and run,
though with no functionality.