Add Log Action
This commit adds a logging action to LuaChat. 'Actions' are pieces
of C++ code (typically classes) that are wrapped by SWIG and made
available to Lua. The 'log' action allows Lua to write messages into
the same logging infrastructure as the main application (i.e. the
spdlog sinks). In addition, this commit includes the CMake code to
run SWIG at build time i.e. to generate the C++ to Lua wrapper.