Add Talk Action
Add a 'talk' action to allow one instance of LuaChat to 'talk' to
another. This action provides the underlying message transport (via
TCP), which permits LuaChat to send strings of text across the network.

The action uses the excellent Asio (v 1.12.2) to provide asynchronous
TCP handling. See the following URL for details:

http://think-async.com/Asio/asio-1.12.2/doc/