Add a Lua Manager
The Lua Manager is responsible for loading and running Lua behaviours.
Each behaviour should include a Lua table (typically located at the
bottom of the file) to provide meta-data to the Lua Manager. This table
should be returned by the script so that it will be available to the
C++ code when the behaviour is loaded. An example is below:

local behaviour = {
    name = "LuaChat",
    description = "A Lua behaviour to allow users to chat over TCP",
    entry_point = main
  }

return behaviour

Further fields can be added as required (e.g. versioning info).