Add Timer Action
Add a 'timer' action to allow Lua to execute blocking and non-blocking
waits. This action wraps the timer support in ASIO and allows Lua to
poll for timer expiration events. In particular, this action will allow
a coroutine scheduler to be implemented (in a subsequent commit) that
uses a blocking wait to prevent LuaChat from consuming 100% of the CPU.