Implement LuaChat Behaviour
This commit implements the 'full' LuaChat behaviour. This includes
corutines for sending and receiving messages plus a dispatcher that
runs the individual coroutines. There are several points of interest:

i)  The dispatcher code is a simplified version of the dispatcher
    presented in section 9.4 of Programming in Lua:

    https://www.lua.org/pil/9.4.html

    Note that the addition of a blocking wait is to prevent LuaChat
    from always requesting the processor (and so consuming 100% of
    the CPU).

ii) The use of rpoll in the sender coroutine is necessary to prevent
    the sender from blocking indefinitely (i.e. with io.read). If
    io.read is called regardless, then the sender coroutine will block
    until the user enters a string. Thus, messages sent by the peer will
    not be displayed until io.read returns.